+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/api/BaseHttpService.scala
@@ -2,9 +2,10 @@ package pl.touk.nussknacker.ui.api
 
 import cats.data.EitherT
 import pl.touk.nussknacker.security.AuthCredentials
-import pl.touk.nussknacker.ui.api.BaseHttpService.{CustomAuthorizationError, LogicResult, NoRequirementServerEndpoint}
+import pl.touk.nussknacker.ui.api.BaseHttpService.{CustomAuthorizationError, HttpServiceServerEndpoint}
 import pl.touk.nussknacker.ui.security.api._
 import pl.touk.nussknacker.ui.security.api.SecurityError.InsufficientPermission
+import sttp.capabilities.pekko.PekkoStreams
 import sttp.tapir.server.{PartialServerEndpoint, ServerEndpoint}
 
 import java.util.concurrent.atomic.AtomicReference
@@ -14,15 +15,16 @@ abstract class BaseHttpService(
     authManager: AuthManager
 )(implicit executionContext: ExecutionContext) {
 
-  private val httpServiceSupport = new HttpServiceSupport(authManager)
+  // the discussion about this approach can be found here: https://github.com/TouK/nussknacker/pull/4685#discussion_r1329794444
+  type LogicResult[BUSINESS_ERROR, RESULT] = Either[Either[BUSINESS_ERROR, SecurityError], RESULT]
 
-  private val allServerEndpoints = new AtomicReference(List.empty[NoRequirementServerEndpoint])
+  private val allServerEndpoints = new AtomicReference(List.empty[HttpServiceServerEndpoint])
 
-  protected def expose(serverEndpoint: NoRequirementServerEndpoint): Unit = {
+  protected def expose(serverEndpoint: HttpServiceServerEndpoint): Unit = {
     expose(List(serverEndpoint))
   }
 
-  protected def expose(serverEndpoints: List[NoRequirementServerEndpoint]): Unit = {
+  protected def expose(serverEndpoints: List[HttpServiceServerEndpoint]): Unit = {
     allServerEndpoints
       .accumulateAndGet(
         serverEndpoints,
@@ -30,25 +32,44 @@ abstract class BaseHttpService(
       )
   }
 
-  protected def expose(when: => Boolean)(serverEndpoint: NoRequirementServerEndpoint): Unit = {
+  protected def expose(when: => Boolean)(serverEndpoint: HttpServiceServerEndpoint): Unit = {
     if (when) expose(serverEndpoint)
   }
 
-  def serverEndpoints: List[NoRequirementServerEndpoint] = allServerEndpoints.get()
+  def serverEndpoints: List[HttpServiceServerEndpoint] = allServerEndpoints.get()
 
   protected def authorizeAdminUser[BUSINESS_ERROR](
       credentials: AuthCredentials
-  ): Future[LogicResult[BUSINESS_ERROR, LoggedUser]] = httpServiceSupport.authorizeAdminUser(credentials)
+  ): Future[LogicResult[BUSINESS_ERROR, LoggedUser]] = {
+    authorizeKnownUser[BUSINESS_ERROR](credentials)
+      .map {
+        case right @ Right(AdminUser(_, _)) => right
+        case Right(_: CommonUser)           => securityError(InsufficientPermission)
+        case Right(_: ImpersonatedUser)     => securityError(InsufficientPermission)
+        case error @ Left(_)                => error
+      }
+  }
 
   protected def authorizeKnownUser[BUSINESS_ERROR](
       credentials: AuthCredentials
-  ): Future[LogicResult[BUSINESS_ERROR, LoggedUser]] = httpServiceSupport.authorizeKnownUser(credentials)
+  ): Future[LogicResult[BUSINESS_ERROR, LoggedUser]] = {
+    authManager
+      .authenticate(credentials)
+      .map {
+        case Left(authenticationError) => securityError(authenticationError)
+        case Right(authenticatedUser) =>
+          authManager.authorize(authenticatedUser) match {
+            case Right(loggedUser)        => success(loggedUser)
+            case Left(authorizationError) => securityError(authorizationError)
+          }
+      }
+  }
 
-  protected def success[RESULT](value: RESULT) = httpServiceSupport.success(value)
+  protected def success[RESULT](value: RESULT) = Right(value)
 
-  protected def businessError[BUSINESS_ERROR](error: BUSINESS_ERROR) = httpServiceSupport.businessError(error)
+  protected def businessError[BUSINESS_ERROR](error: BUSINESS_ERROR) = Left(Left(error))
 
-  protected def securityError[SE <: SecurityError](error: SE) = httpServiceSupport.securityError(error)
+  protected def securityError[SE <: SecurityError](error: SE) = Left(Right(error))
 
   private type PartialEndpoint[INPUT, OUTPUT, BUSINESS_ERROR, -R] =
     PartialServerEndpoint[_, LoggedUser, INPUT, Either[BUSINESS_ERROR, SecurityError], OUTPUT, R, Future]
@@ -84,53 +105,9 @@ abstract class BaseHttpService(
 }
 
 object BaseHttpService {
-  // the discussion about this approach can be found here: https://github.com/TouK/nussknacker/pull/4685#discussion_r1329794444
-  type LogicResult[BUSINESS_ERROR, RESULT] = Either[Either[BUSINESS_ERROR, SecurityError], RESULT]
 
-  // we assume that our endpoints have no special requirements (in the Tapir sense)
-  type NoRequirementServerEndpoint = ServerEndpoint[Any, Future]
+  type HttpServiceServerEndpoint = ServerEndpoint[Any with PekkoStreams, Future]
 
   // it's marker interface which simplifies error handling when serverLogicEitherT is used
   trait CustomAuthorizationError
-}
-
-class HttpServiceSupport(
-    authManager: AuthManager
-)(implicit executionContext: ExecutionContext) {
-
-  def authorizeAdminUser[BUSINESS_ERROR](
-      credentials: AuthCredentials
-  ): Future[LogicResult[BUSINESS_ERROR, LoggedUser]] = {
-    authorizeKnownUser[BUSINESS_ERROR](credentials)
-      .map {
-        case right @ Right(AdminUser(_, _)) => right
-        case Right(_: CommonUser)           => securityError(InsufficientPermission)
-        case Right(_: ImpersonatedUser)     => securityError(InsufficientPermission)
-        case error @ Left(_)                => error
-      }
-  }
-
-  def authorizeKnownUser[BUSINESS_ERROR](
-      credentials: AuthCredentials
-  ): Future[LogicResult[BUSINESS_ERROR, LoggedUser]] = {
-    authManager
-      .authenticate(credentials)
-      .map {
-        case Left(authenticationError) => securityError(authenticationError)
-        case Right(authenticatedUser) =>
-          authManager.authorize(authenticatedUser) match {
-            case Right(loggedUser)        => success(loggedUser)
-            case Left(authorizationError) => securityError(authorizationError)
-          }
-      }
-  }
-
-  def success[RESULT](value: RESULT): Right[Nothing, RESULT] = Right(value)
-
-  def businessError[BUSINESS_ERROR](error: BUSINESS_ERROR): Left[Left[BUSINESS_ERROR, Nothing], Nothing] = Left(
-    Left(error)
-  )
-
-  def securityError[SE <: SecurityError](error: SE): Left[Right[Nothing, SE], Nothing] = Left(Right(error))
-
 }
