+++ designer/server/src/main/scala/pl/touk/nussknacker/ui/server/TapirHttpServiceFactory.scala
@@ -3,7 +3,7 @@ package pl.touk.nussknacker.ui.server
 import pl.touk.nussknacker.engine.util.ResourceLoader
 import pl.touk.nussknacker.ui.api._
 import pl.touk.nussknacker.ui.config.DesignerConfig
-import pl.touk.nussknacker.ui.customhttpservice.{PekkoCustomHttpServiceProvider, TapirCustomHttpServiceProvider}
+import pl.touk.nussknacker.ui.customhttpservice.TapirCustomHttpServiceProvider
 import pl.touk.nussknacker.ui.factory.{DomainServices, InfrastructureServices}
 import pl.touk.nussknacker.ui.migrations.{MigrationApiAdapterService, MigrationService}
 import pl.touk.nussknacker.ui.notifications.NotificationServiceImpl
@@ -218,9 +218,10 @@ object TapirHttpServiceFactory {
       )
     }
 
+    val httpServiceExposingTapirCustomProviders =
+      new HttpServiceExposingTapirCustomProviders(customHttpServiceProviders, authManager)
+
     new NuDesignerExposedApiHttpService(
-      customHttpServiceProviders,
-      authManager,
       appApiHttpService,
       componentsApiHttpService,
       dictApiHttpService,
@@ -235,7 +236,8 @@ object TapirHttpServiceFactory {
       scenarioParametersHttpService,
       userApiHttpService,
       versionControlApiHttpService,
-      statisticsApiHttpService
+      statisticsApiHttpService,
+      httpServiceExposingTapirCustomProviders,
     )
   }
